chore: move styles to screens and components directories
