+++ src/styles/components/FeaturedCategories.ts
