+++ src/styles/components/ImageCarousel.ts
