+++ src/styles/components/PDPDetails.ts
