+++ src/styles/components/featuredProducts.ts
