+++ src/styles/screens/Categories.ts
