+++ src/styles/screens/Scan.ts
